perplexity-ios: the Dynamic Island, four Live Activity states
Boards 11-14: voice mode and reasoning, each compact and expanded. They diff
at 1.48, 1.84, 1.52 and 1.51, the four best in the folder; mean over the
fourteen is 3.26, and the ten onboarding screens are unmoved at 3.92.

Their captures are 881x2000, not @3x -- an 881x1910 screen under a 90px
strip, so 2.24173 px per pt, which the height agrees with at 1910/852.
scale() reads that off the capture's number, and each island probe row
carries its own pt so refkit batch overrides the command line. refkit shoot's
--scale was type=int and would not take 2.24173; it is a float now.

The status-bar rule says nothing of the capture's own top 54pt is carried
over, an expanded island included. These four are the exception it names,
because the Live Activity is the subject of the board: statusbar() still
draws the 9:41 clock and the signal group, and the island goes over the
template's plain one. That costs 19 to 33 levels in the top 54pt, almost all
of it the clock, and on p11 alone leaves the fourth cellular bar showing past
the narrower pill.

The expanded island is cut at its own box rather than the frame, so the
board's own 43.75pt corner clips the starfield and the shadow under it is
CSS, solved from a Gaussian at sigma 16.5pt. The compact glyph's glow is
alpha = A*max(0, 1 - d/16.5), least-squares fitted to the annulus below each
glyph where the pill is otherwise black: A 0.41 on p11 and 0.73 on p13, one
colour between them. A closest-side radial in a 41px box matched its own
model and fell far slower than the capture, at probe delta 18 against 6.

77 tokens, 67 probes -- 43 colour at a mean worst-channel 2.9, 22 box at mean
|dw| 0.68 and |dh| 0.54. The README records what is left of the compact pill:
the brand mark's own file carries 17 to 35% less ink than the capture below
about 20pt while its ink box matches exactly, and no single stroke fits all
four sizes the folder places it at, so it ships unthickened.

+++ tools/refkit.py
@@ -1348,7 +1348,7 @@ def _parser():
     t = s.add_parser("shoot"); t.set_defaults(fn=cmd_shoot)
     t.add_argument("html", nargs="+"); t.add_argument("-o", "--out", required=True)
     t.add_argument("--w", type=int, default=478); t.add_argument("--h", type=int, default=980)
-    t.add_argument("--scale", type=int, default=2)
+    t.add_argument("--scale", type=float, default=2)
     t.add_argument("--crop-phone", action="store_true",
                    help="cut the 393x852 screen out of the frame, corners masked "
                         "to the 52pt radius, ready to diff")
